+++ verifier/src/main/java/com/aws/jverify/verifier/dafny/Serializer.java
@@ -1,9 +1,8 @@
-package com.aws.jverify.verifier;
+package com.aws.jverify.verifier.dafny;
 
 import com.aws.jverify.generated.BoundVar;
 import com.aws.jverify.generated.Expression;
 import com.aws.jverify.generated.FrameExpression;
-import com.aws.jverify.verifier.dafny.Encoder;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.lang.reflect.*;
