+++ verifier/src/main/java/com/aws/jverify/verifier/dafny/StatusMessage.java
@@ -1,6 +1,5 @@
-package com.aws.jverify.verifier;
+package com.aws.jverify.verifier.dafny;
 
-import com.aws.jverify.verifier.dafny.DafnyOutput;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public final class StatusMessage extends DafnyOutput {
